+++ main_window.py
@@ -1,437 +1,15 @@
-import sys
 from datetime import datetime
 from typing import List, Optional
-from dataclasses import dataclass
-from enum import Enum
-
 from PySide6 import QtCore, QtGui, QtWidgets
-import zmq
-import base64
-import numpy as np
-import cv2
-from std_msgs.msg import Empty
+import base64, zmq, numpy as np, cv2
 
-# ---------------------------
-# Configuration & Constants
-# ---------------------------
-@dataclass
-class ButtonConfig:
-    """버튼 설정 클래스"""
-    flag_name: str
-    button_name: str
-    text_idle: str
-    text_active: str
-    button_type: str = 'detection'  # 'detection' or 'robot'
+from .config import DEFAULT_ROS_TOPICS, UI_STYLES, MissionType
+from .state import StateManager
+from .camera_label import CameraLabel
+from .zmq_subscriber import ZmqSubscriberThread
+from .ros2_subscriber import Ros2SubscriberThread
 
-@dataclass
-class UIStyles:
-    """UI 스타일 상수"""
-    DET_IDLE: str = "color: white; background: darkred;"
-    DET_ACTIVE: str = "color: black; background: yellow;"
-    ROBOT_IDLE: str = "color: black; background: white; border:1px solid #8b1f1f; padding:6px;"
-    ROBOT_ACTIVE: str = "color: white; background: #2ecc71; border:1px solid #1e8449; padding:6px;"
 
-class MissionType(Enum):
-    MISSION1 = 0
-    MISSION2 = 1  
-    MISSION3 = 2
-    MISSION4 = 3
-
-# 설정 집중화
-UI_STYLES = UIStyles()
-
-BUTTON_CONFIGS = {
-    'SOS': ButtonConfig('sos_flag', 'btn_m1_sos', 'SOS: NO', 'SOS: YES'),
-    'Button': ButtonConfig('button_flag', 'btn_m3_button', 'BUTTON: NO', 'BUTTON: YES'),
-    'Fire': ButtonConfig('fire_flag', 'btn_m3_fire', 'FIRE: NO', 'FIRE: YES'),
-    'Door': ButtonConfig('door_flag', 'btn_m3_door', 'DOOR: NO', 'DOOR: YES'),
-    'Safebox': ButtonConfig('safebox_flag', 'btn_m4_safebox', 'SAFEBOX: NO', 'SAFEBOX: YES'),
-    'Human': ButtonConfig('human_flag', 'btn_m4_human', 'HUMAN: NO', 'HUMAN: YES'),
-    'Finish': ButtonConfig('finish_flag', 'btn_m4_finish', 'FINISH: NO', 'FINISH: YES'),
-    'OK': ButtonConfig('robot_ok_flag', 'btn_ok', 'OK', 'OK', 'robot'),
-    'Open': ButtonConfig('robot_open_flag', 'btn_m3_open', 'Open', 'Open', 'robot'),
-    'Pick': ButtonConfig('robot_pick_flag', 'btn_m4_pick', 'Pick', 'Pick', 'robot'),
-}
-
-DEFAULT_ROS_TOPICS = {
-    'ok': '/robot/ok',
-    'open': '/robot/open',
-    'pick': '/robot/pick'
-}
-
-# ---------------------------
-# 공통 베이스 클래스
-# ---------------------------
-class BaseSubscriberThread(QtCore.QThread):
-    """공통 구독자 스레드 베이스 클래스"""
-    text_received = QtCore.Signal(str)
-    
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._stop = False
-    
-    def stop(self):
-        self._stop = True
-    
-    def emit_log(self, message: str):
-        """로그 메시지 발송"""
-        self.text_received.emit(message)
-
-# ---------------------------
-# ZMQ 구독자
-# ---------------------------
-class ZmqSubscriberThread(BaseSubscriberThread):
-    frame_received = QtCore.Signal(object)
-    
-    def __init__(self, endpoints: List[str], context=None, parent=None):
-        super().__init__(parent)
-        self.endpoints = list(endpoints)
-        self._ctx_owned = context is None
-        self.context = context or zmq.Context(io_threads=1)
-        
-    def _decode_frame(self, payload_bytes: bytes) -> Optional[np.ndarray]:
-        """프레임 디코딩 최적화"""
-        if not payload_bytes:
-            return None
-            
-        # Base64 디코딩 시도
-        try:
-            jpg_bytes = base64.b64decode(payload_bytes, validate=False)
-            np_arr = np.frombuffer(jpg_bytes, dtype=np.uint8)
-            frame = cv2.imdecode(np_arr, cv2.IMREAD_COLOR)
-            if frame is not None:
-                return frame
-        except Exception:
-            pass
-        
-        # 직접 바이너리 디코딩 시도
-        try:
-            np_arr = np.frombuffer(payload_bytes, dtype=np.uint8)
-            return cv2.imdecode(np_arr, cv2.IMREAD_COLOR)
-        except Exception as e:
-            self.emit_log(f"[ZMQ] Frame decode failed: {e}")
-            return None
-    
-    def run(self):
-        sockets = []
-        poller = zmq.Poller()
-        
-        try:
-            for ep in self.endpoints:   # 소켓 초기화
-                try:
-                    sock = self.context.socket(zmq.SUB)
-                    sock.setsockopt(zmq.RCVHWM, 50)
-                    sock.linger = 0
-                    sock.connect(ep)
-                    sock.setsockopt_string(zmq.SUBSCRIBE, "")
-                    poller.register(sock, zmq.POLLIN)
-                    sock.setsockopt(zmq.CONFLATE, 1)  # 최신 1개만 유지 (옵션)
-                    sock.setsockopt(zmq.RCVHWM, 1)    # HWM 최소화
-                    sockets.append(sock)
-                    self.emit_log(f"[ZMQ] Connected to {ep}")
-                except Exception as e:
-                    self.emit_log(f"[ZMQ] Connection error {ep}: {e}")
-                    continue
-            
-            while not self._stop:     # 메인 루프
-                try:
-                    events = dict(poller.poll(timeout=200))
-                    if not events:
-                        continue
-                    
-                    for sock in events:
-                        if events[sock] & zmq.POLLIN:
-                            try:
-                                msg = sock.recv_multipart(flags=zmq.NOBLOCK)
-                                payload = msg[-1] if msg else b""
-                                
-                                frame = self._decode_frame(payload)
-                                if frame is not None:
-                                    self.frame_received.emit(("camera", frame))
-                                    
-                            except zmq.Again:
-                                continue
-                            except Exception as e:
-                                self.emit_log(f"[ZMQ] Receive error: {e}")
-                                
-                except Exception as e:
-                    self.emit_log(f"[ZMQ] Poll error: {e}")
-                    
-        finally:
-            self._cleanup_sockets(sockets, poller)
-            self.emit_log("[ZMQ] Thread terminated")
-    
-    def _cleanup_sockets(self, sockets: List, poller: zmq.Poller):
-        """소켓 정리"""
-        for sock in sockets:
-            try:
-                poller.unregister(sock)
-                sock.close(0)
-            except:
-                pass
-        
-        if self._ctx_owned:
-            try:
-                self.context.term()
-            except:
-                pass
-
-class Ros2SubscriberThread(BaseSubscriberThread):
-    # 탐지 시그널들
-    sos_detected        = QtCore.Signal()
-    button_detected     = QtCore.Signal()
-    fire_detected       = QtCore.Signal()
-    door_detected       = QtCore.Signal()
-    safebox_detected    = QtCore.Signal()
-    safebox2_detected   = QtCore.Signal()
-    human_tick          = QtCore.Signal()
-    finish_detected     = QtCore.Signal()
-    
-    # 로봇 상태 시그널들
-    robot_ok            = QtCore.Signal()
-    robot_open          = QtCore.Signal()
-    robot_pick          = QtCore.Signal()
-    robot_button_end    = QtCore.Signal()
-    robot_open_door_end = QtCore.Signal()
-    pick_place_end      = QtCore.Signal()
-
-    
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._node = None
-        self._subs = []
-        
-    def _create_subscriptions(self):
-        """구독 생성 """
-        # 탐지 토픽들
-        detection_topics = {
-            # mission 1
-            '/object_detection/SOS': self.sos_detected,
-            # mission 3
-            '/object_detection/button': self.button_detected,
-            '/object_detection/fire/on': self.fire_detected,
-            '/object_detection/fire/off': self.fire_detected,
-            '/object_detection/door': self.door_detected,
-            #mission 4
-            '/object_detection/safebox': self.safebox_detected,
-            '/object_detection/safebox_2': self.safebox2_detected,
-            '/object_detection/human/red': self.human_tick,
-            '/object_detection/human/blue': self.human_tick,
-            '/object_detection/finish': self.finish_detected,
-        }
-        # ----------- 로봇 상태 토픽들 -----------------
-        robot_topics = {
-            '/robot/ok': self.robot_ok,
-            '/robot/open': self.robot_open,
-            '/robot/pick': self.robot_pick,
-        }
-        # ---------- mission3 로봇암 완료 토픽 ----------
-        robot_arm_topics = {
-            '/robot_arm/button_end'   : self.robot_button_end,
-            '/robot_arm/open_door_end': self.robot_open_door_end,
-            '/robot_arm/pick_and_place_end': self.pick_place_end,
-        }
-
-        # 모든 토픽 구독
-        all_topics = {**detection_topics, **robot_topics, **robot_arm_topics}
-        
-        for topic, signal in all_topics.items():
-            try:
-                sub = self._node.create_subscription(
-                    Empty, topic,
-                    lambda _msg, sig=signal: sig.emit(),
-                    10
-                )
-                self._subs.append(sub)
-                self.emit_log(f"[ROS2] Subscribed to {topic}")
-            except Exception as e:
-                self.emit_log(f"[ROS2] Subscription error {topic}: {e}")
-
-    def run(self):
-        try:
-            import rclpy
-            from std_msgs.msg import Empty
-        except ImportError as e:
-            self.emit_log(f"[ROS2] Import error: {e}")
-            return
-        
-        try:
-            # ROS2 초기화
-            try:
-                rclpy.init(args=None)
-            except:
-                pass  # 이미 초기화된 경우
-            
-            self._node = rclpy.create_node('ui_subscriber')
-            self._create_subscriptions()
-            
-            # 메인 루프
-            while not self._stop and rclpy.ok():
-                try:
-                    rclpy.spin_once(self._node, timeout_sec=0.1)
-                except Exception as e:
-                    self.emit_log(f"[ROS2] Spin error: {e}")
-                    
-        except Exception as e:
-            self.emit_log(f"[ROS2] Runtime error: {e}")
-        finally:
-            self._cleanup()
-            self.emit_log("[ROS2] Thread terminated")
-    
-    def _cleanup(self):
-        """리소스 정리"""
-        if self._node:
-            try:
-                self._node.destroy_node()
-            except:
-                pass
-        
-        try:
-            import rclpy
-            rclpy.shutdown()
-        except:
-            pass
-
-# ---------------------------
-# 최적화된 카메라 레이블
-# ---------------------------
-class CameraLabel(QtWidgets.QLabel):
-    def __init__(self, placeholder_text="No Image", parent=None):
-        super().__init__(parent)
-        self.setAlignment(QtCore.Qt.AlignCenter)
-        self.setText(placeholder_text)
-        self.setMinimumSize(640, 480)
-        self.setStyleSheet("background-color: #111; color: #eee;")
-        self._current_pixmap = None
-        self._last_size = self.size()
-        
-    def show_frame(self, frame_bgr: np.ndarray):
-        """프레임 표시 최적화"""
-        if not self._is_valid_frame(frame_bgr):
-            self._show_no_image()
-            return
-        
-        try:
-            qimg = self._convert_frame_to_qimage(frame_bgr)
-            if qimg.isNull():
-                self._show_no_image()
-                return
-            
-            pixmap = QtGui.QPixmap.fromImage(qimg)
-            self._current_pixmap = pixmap
-            self._update_display()
-            
-        except Exception as e:
-            print(f"[CameraLabel] Error: {e}", file=sys.stderr)
-            self._show_no_image()
-    
-    def _is_valid_frame(self, frame: np.ndarray) -> bool:
-        """프레임 유효성 검사"""
-        return (frame is not None and 
-                isinstance(frame, np.ndarray) and 
-                frame.size > 0)
-    
-    def _convert_frame_to_qimage(self, frame_bgr: np.ndarray) -> QtGui.QImage:
-        """BGR 프레임을 QImage로 변환"""
-        rgb = cv2.cvtColor(frame_bgr, cv2.COLOR_BGR2RGB)
-        rgb = np.ascontiguousarray(rgb)
-        h, w, ch = rgb.shape
-        bytes_per_line = ch * w
-        return QtGui.QImage(rgb.data, w, h, bytes_per_line, QtGui.QImage.Format_RGB888).copy()
-    
-    def _show_no_image(self):
-        """이미지 없음 표시"""
-        self.clear()
-        self.setText("No Image")
-        self._current_pixmap = None
-    
-    def _update_display(self):
-        if self._current_pixmap and not self._current_pixmap.isNull():
-            if self.size() != self._last_size:
-                self._scaled_pixmap = self._current_pixmap.scaled(
-                    self.size(),
-                    QtCore.Qt.KeepAspectRatio,
-                    QtCore.Qt.SmoothTransformation)
-                self._last_size = self.size()
-            self.setPixmap(getattr(self, "_scaled_pixmap", self._current_pixmap))
-    
-    def resizeEvent(self, event):
-        """리사이즈 이벤트 처리"""
-        self._last_size = QtCore.QSize(0,0) # 다음 _update_display에서 재스케일 강제
-        self._update_display()
-        super().resizeEvent(event)
-
-# ---------------------------
-# UI 상태 관리자
-# ---------------------------
-class StateManager:
-    """UI 상태 관리 클래스"""
-    
-    def __init__(self):
-        # 감지 플래그들
-        self.detection_flags = {
-            'sos_flag': False,
-            'button_flag': False, 
-            'fire_flag': False,
-            'door_flag': False,
-            'safebox_flag': False,
-            'safebox2_flag': False,
-            'human_flag': False,
-            'finish_flag': False
-        }
-        
-        # 로봇 상태 플래그들
-        self.robot_flags = {
-            'robot_ok_flag': False,
-            'robot_open_flag': False,
-            'robot_pick_flag': False,
-            'robot_pick_place_flag': False,
-            'robot_fire_done_flag': False,
-            'robot_human_done_flag': False,
-        }
-        
-        # 카운터들
-        self.counters = {
-            'fire_count': 0,
-            'human_count': 0
-        }
-    
-    def set_flag(self, flag_name: str, value: bool):
-        """플래그 설정"""
-        if flag_name in self.detection_flags:
-            self.detection_flags[flag_name] = value
-        elif flag_name in self.robot_flags:
-            self.robot_flags[flag_name] = value
-    
-    def get_flag(self, flag_name: str) -> bool:
-        """플래그 조회"""
-        return (self.detection_flags.get(flag_name, False) or 
-                self.robot_flags.get(flag_name, False))
-    
-    def increment_counter(self, counter_name: str) -> int:
-        """카운터 증가"""
-        if counter_name in self.counters:
-            self.counters[counter_name] += 1
-            return self.counters[counter_name]
-        return 0
-    
-    def reset_counter(self, counter_name: str):
-        """카운터 리셋"""
-        if counter_name in self.counters:
-            self.counters[counter_name] = 0
-    
-    def reset_all(self):
-        """모든 상태 리셋"""
-        for key in self.detection_flags:
-            self.detection_flags[key] = False
-        for key in self.robot_flags:
-            self.robot_flags[key] = False
-        for key in self.counters:
-            self.counters[key] = 0
-
-# ---------------------------
-# 메인 윈도우 (최적화)
-# ---------------------------
 class MainWindow(QtWidgets.QMainWindow):
     def __init__(self, zmq_endpoints: List[str], ros_topics=None):
         super().__init__()
@@ -606,7 +184,7 @@ class MainWindow(QtWidgets.QMainWindow):
         
         # 버튼들
         self.btn_m3_button = self._create_button("BUTTON: NO")
-        self.btn_m3_fire = self._create_button("FIRE: (0/3)")
+        self.btn_m3_fire = self._create_button("FIRE: (0/4)")
         self.btn_m3_door = self._create_button("DOOR: NO")
         
         # 이벤트 연결
@@ -802,7 +380,7 @@ class MainWindow(QtWidgets.QMainWindow):
         """mission3 불 플래그/카운터 리셋"""
         self.state.set_flag('fire_flag', False)
         self.state.reset_counter('fire_count')
-        self.btn_m3_fire.setText("FIRE: (0/3)")
+        self.btn_m3_fire.setText("FIRE: (0/4)")
         self.btn_m3_fire.setStyleSheet(UI_STYLES.DET_IDLE)
         # fire 완료 로봇 박스도 함께 리셋할지 
         self.state.set_flag('robot_fire_done_flag', False)
@@ -879,7 +457,7 @@ class MainWindow(QtWidgets.QMainWindow):
             self.main_cam.show_frame(frame)
         finally:
             self._frame_pending = False
-    
+
     # -------------------------------------------------------------
     # ROS2 콜백 함수들, flag를 설정 
     # -------------------------------------------------------------
@@ -905,23 +483,25 @@ class MainWindow(QtWidgets.QMainWindow):
     
     @QtCore.Slot()
     def _on_fire_detected(self):
-        """불 감지: on/off 아무거나 오면 (n/3) 증가, 노랑, 3/3이면 로봇 Fire 초록"""
-        # 카운터 +1 (최대 3으로 캡)
-        n = min(3, self.state.increment_counter('fire_count'))
-
+        """불 감지: on/off 아무거나 오면 (n/3) 증가, 최대 3개 까지 증가"""
+        n = self.state.increment_counter('fire_count')
+        if n > 4:
+            n = 4
+            self.state.counters['fire_count'] = 4
+        
         # 감지 버튼(빨간→노랑) & 텍스트 갱신
         if not self.state.get_flag('fire_flag'):
             self.state.set_flag('fire_flag', True)
-            self.btn_m3_fire.setStyleSheet(UI_STYLES.DET_ACTIVE)  # 노랑
-            self._append_log("[ROS2] 불 인식 완료! Fire detected!")
-        self.btn_m3_fire.setText(f"FIRE: ({n}/3)")
+            self.btn_m3_fire.setStyleSheet(UI_STYLES.DET_ACTIVE)
+        self.btn_m3_fire.setText(f"FIRE: ({n}/4)")
+        self._append_log(f"[ROS2] 불 인식중... -> {n}/4")
 
         # 3/3 도달 시 로봇 Fire 박스 초록
-        if n >= 3 and not self.state.get_flag('robot_fire_done_flag'):
+        if n >= 4 and not self.state.get_flag('robot_fire_done_flag'):
             self.state.set_flag('robot_fire_done_flag', True)
             if hasattr(self, 'btn_m3_fire_done'):
-                self.btn_m3_fire_done.setStyleSheet(UI_STYLES.ROBOT_ACTIVE)  # 초록
-            self._append_log("[ROS2] 불 3/3개 모두 인식 완료! -> Robot Fire Done!")
+                self.btn_m3_fire_done.setStyleSheet(UI_STYLES.ROBOT_ACTIVE)
+            self._append_log("[ROS2] 불 인식 완료!(4/4)-> Robot Fire Done!")
     
     @QtCore.Slot()
     def _on_door_detected(self):
@@ -988,7 +568,7 @@ class MainWindow(QtWidgets.QMainWindow):
             self.state.set_flag('finish_flag', True)
             self.btn_m4_finish.setText("FINISH: YES")
             self.btn_m4_finish.setStyleSheet(UI_STYLES.DET_ACTIVE)
-            self._append_log("[ROS2] Finish라인 인식! Finish detected!")
+            self._append_log("[ROS2] Finish라인 인식 완료! Finish detected!")
 
     @QtCore.Slot()
     def _on_robot_button_end(self):
@@ -1134,29 +714,3 @@ class MainWindow(QtWidgets.QMainWindow):
     def _cleanup(self):
         """리소스 정리"""
         self._stop_threads()
-
-# ---------------------------
-# main
-# ---------------------------
-def main():
-    """메인 함수"""
-    # 기본 엔드포인트 설정
-    endpoints = [
-        "tcp://127.0.0.1:5555",  # 메인 카메라
-        # "tcp://127.0.0.1:5556",  # 보급상자 감지
-        # "tcp://127.0.0.1:5558",  # 사람 감지
-        # "tcp://127.0.0.1:5559",  # 완료 감지
-    ]
-    
-    app = QtWidgets.QApplication(sys.argv)
-    window = MainWindow(endpoints)
-    window.show()
-    
-    try:
-        sys.exit(app.exec())
-    except KeyboardInterrupt:
-        print("\nApplication interrupted by user")
-        sys.exit(0)
-
-if __name__ == "__main__":
-    main()
